Update posts and add dotfiles

+++ build.sh
@@ -1,10 +1,14 @@
 # /bin/bash
 
 # Build local iste
-JEKYLL_ENV=production bundle exec jekyll build --config _config.yml,_config.local.yml
-# Delete old file and copy over new ones (local build)
-sudo rm -rf /var/www/blog.pradosh.dev
-sudo cp -r _site /var/www/blog.pradosh.dev
-# Build and deploy to firebase
-JEKYLL_ENV=production bundle exec jekyll build --config _config.yml,_config.firebase.yml
-firebase deploy
+if [[ $1 == "local" || $1 == "all" ]]; then
+  JEKYLL_ENV=production bundle exec jekyll build --config _config.yml,_config.local.yml
+  # Delete old file and copy over new ones (local build)
+  sudo rm -rf /var/www/blog.pradosh.dev
+  sudo cp -r _site /var/www/blog.pradosh.dev
+fi
+if [[ $1 == "firebase" || $1 == "all" ]]; then
+  # Build and deploy to firebase
+  JEKYLL_ENV=production bundle exec jekyll build --config _config.yml,_config.firebase.yml
+  firebase deploy
+fi
